pyproject.toml
[project]
name = "startup-idea-validator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.12.0",
    "google-genai>=1.0.0",
    "ipython>=9.4.0",
    "langchain-tavily>=0.2.11",
    "nest-asyncio>=1.6.0",
    "streamlit>=1.48.1",
    "tavily-python>=0.7.11",
]
